requirements.txt
playwright>=1.40.0
pyyaml>=6.0.1
apscheduler>=3.10.4
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
jinja2>=3.1.2
pytz
